make currentNum concatenate
I added the functions to make num1 concatenate numbers that you press.
that means that each successive number will be concatenated to the previous numbers.

I found out that num1 was not updating to that value of currentNum.
Because of that I have separated the "switchCurrentNum" into 2 parts:

a flag that switches TRUE to FALSE every time it is run.
a function "updateNum" which is just there to update num1 or num2,
depending on the flag.

this function can be run before updating the display