make operator btns be generated from an object
I removed the HTML operator buttons, and made these be generated from an object.
I also changed the name of the operator container in html.
I used to be called "dynamicBtnFrame" now it is:
"btnFrameOperator".

in javascript,
I placed the container node declarations before other node declarations.
this makes it harder to append to an undeclared node.