make user input be a starting
I decided that the inout for the calculation will be a string.
the calculator would be easier to to scale:

- it will be able to to take a long calculation in the future:
If I already start with evaluating the calculation as a string,
I can keep my code that does calculations the same, when I have long calculations.

It will give less errors.
whatever you type appears on the right, so no need to switch between variables.

I make a new variable called "userInput"
I edited the updateDisplay function, to display the userInput variable.