bugfix, removed infinite loop calculateInOrder
I made another file (tryout.js)
to see if splice could work on an object, which it did.
when, in splice() I changed the removed value from (index-1) to
(index), the infinite loop stopped.

For now I can keep working at least, and try to remove double operators,
after that I can see if there will still be a problem with .splice().

I changed the condition of looping calculateInOrder to 2 things,
when an operator of the current precedence level was found,
check the array again

when no operator was found, and the lowest precedence level was reached,
end loop.